Move computeEnforced to PushPredicateIntoTableScan
