+++ core/trino-main/src/main/java/io/trino/metadata/TableLayoutResult.java
@@ -13,14 +13,9 @@
  */
 package io.trino.metadata;
 
-import com.google.common.collect.ImmutableMap;
 import io.trino.spi.connector.ColumnHandle;
-import io.trino.spi.predicate.Domain;
 import io.trino.spi.predicate.TupleDomain;
 
-import java.util.Map;
-
-import static com.google.common.base.Preconditions.checkArgument;
 import static java.util.Objects.requireNonNull;
 
 @Deprecated
@@ -50,55 +45,5 @@ public class TableLayoutResult
     public TupleDomain<ColumnHandle> getUnenforcedConstraint()
     {
         return unenforcedConstraint;
-    }
-
-    public static TupleDomain<ColumnHandle> computeEnforced(TupleDomain<ColumnHandle> predicate, TupleDomain<ColumnHandle> unenforced)
-    {
-        if (predicate.isNone()) {
-            // If the engine requests that the connector provides a layout with a domain of "none". The connector can have two possible reactions, either:
-            // 1. The connector can provide an empty table layout.
-            //   * There would be no unenforced predicate, i.e., unenforced predicate is TupleDomain.all().
-            //   * The predicate was successfully enforced. Enforced predicate would be same as predicate: TupleDomain.none().
-            // 2. The connector can't/won't.
-            //   * The connector would tell the engine to put a filter on top of the scan, i.e., unenforced predicate is TupleDomain.none().
-            //   * The connector didn't successfully enforce anything. Therefore, enforced predicate would be TupleDomain.all().
-            if (unenforced.isNone()) {
-                return TupleDomain.all();
-            }
-            if (unenforced.isAll()) {
-                return TupleDomain.none();
-            }
-            throw new IllegalArgumentException();
-        }
-
-        // The engine requested the connector provides a layout with a non-none TupleDomain.
-        // A TupleDomain is effectively a list of column-Domain pairs.
-        // The connector is expected enforce the respective domain entirely on none, some, or all of the columns.
-        // 1. When the connector could enforce none of the domains, the unenforced would be equal to predicate;
-        // 2. When the connector could enforce some of the domains, the unenforced would contain a subset of the column-Domain pairs;
-        // 3. When the connector could enforce all of the domains, the unenforced would be TupleDomain.all().
-
-        // In all 3 cases shown above, the unenforced is not TupleDomain.none().
-        checkArgument(!unenforced.isNone());
-
-        Map<ColumnHandle, Domain> predicateDomains = predicate.getDomains().get();
-        Map<ColumnHandle, Domain> unenforcedDomains = unenforced.getDomains().get();
-        ImmutableMap.Builder<ColumnHandle, Domain> enforcedDomainsBuilder = ImmutableMap.builder();
-        for (Map.Entry<ColumnHandle, Domain> entry : predicateDomains.entrySet()) {
-            ColumnHandle predicateColumnHandle = entry.getKey();
-            if (unenforcedDomains.containsKey(predicateColumnHandle)) {
-                checkArgument(
-                        entry.getValue().equals(unenforcedDomains.get(predicateColumnHandle)),
-                        "Enforced tuple domain cannot be determined. The connector is expected to enforce the respective domain entirely on none, some, or all of the column.");
-            }
-            else {
-                enforcedDomainsBuilder.put(predicateColumnHandle, entry.getValue());
-            }
-        }
-        Map<ColumnHandle, Domain> enforcedDomains = enforcedDomainsBuilder.build();
-        checkArgument(
-                enforcedDomains.size() + unenforcedDomains.size() == predicateDomains.size(),
-                "Enforced tuple domain cannot be determined. Connector returned an unenforced TupleDomain that contains columns not in predicate.");
-        return TupleDomain.withColumnDomains(enforcedDomains);
     }
 }
