+++ core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/PushPredicateIntoTableScan.java
@@ -15,6 +15,7 @@ package io.trino.sql.planner.iterative.rule;
 
 import com.google.common.collect.ImmutableBiMap;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import io.trino.Session;
 import io.trino.cost.StatsProvider;
 import io.trino.matching.Capture;
@@ -28,6 +29,7 @@ import io.trino.metadata.TableProperties.TablePartitioning;
 import io.trino.spi.connector.ColumnHandle;
 import io.trino.spi.connector.Constraint;
 import io.trino.spi.connector.ConstraintApplicationResult;
+import io.trino.spi.predicate.Domain;
 import io.trino.spi.predicate.TupleDomain;
 import io.trino.sql.PlannerContext;
 import io.trino.sql.planner.DomainTranslator;
@@ -46,11 +48,11 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 
+import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Verify.verify;
 import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static io.trino.SystemSessionProperties.isAllowPushdownIntoConnectors;
 import static io.trino.matching.Capture.newCapture;
-import static io.trino.metadata.TableLayoutResult.computeEnforced;
 import static io.trino.sql.ExpressionUtils.combineConjuncts;
 import static io.trino.sql.ExpressionUtils.filterDeterministicConjuncts;
 import static io.trino.sql.ExpressionUtils.filterNonDeterministicConjuncts;
@@ -331,5 +333,55 @@ public class PushPredicateIntoTableScan
         expression = SimplifyExpressions.rewrite(expression, session, symbolAllocator, plannerContext, typeAnalyzer);
 
         return expression;
+    }
+
+    public static TupleDomain<ColumnHandle> computeEnforced(TupleDomain<ColumnHandle> predicate, TupleDomain<ColumnHandle> unenforced)
+    {
+        if (predicate.isNone()) {
+            // If the engine requests that the connector provides a layout with a domain of "none". The connector can have two possible reactions, either:
+            // 1. The connector can provide an empty table layout.
+            //   * There would be no unenforced predicate, i.e., unenforced predicate is TupleDomain.all().
+            //   * The predicate was successfully enforced. Enforced predicate would be same as predicate: TupleDomain.none().
+            // 2. The connector can't/won't.
+            //   * The connector would tell the engine to put a filter on top of the scan, i.e., unenforced predicate is TupleDomain.none().
+            //   * The connector didn't successfully enforce anything. Therefore, enforced predicate would be TupleDomain.all().
+            if (unenforced.isNone()) {
+                return TupleDomain.all();
+            }
+            if (unenforced.isAll()) {
+                return TupleDomain.none();
+            }
+            throw new IllegalArgumentException();
+        }
+
+        // The engine requested the connector provides a layout with a non-none TupleDomain.
+        // A TupleDomain is effectively a list of column-Domain pairs.
+        // The connector is expected enforce the respective domain entirely on none, some, or all of the columns.
+        // 1. When the connector could enforce none of the domains, the unenforced would be equal to predicate;
+        // 2. When the connector could enforce some of the domains, the unenforced would contain a subset of the column-Domain pairs;
+        // 3. When the connector could enforce all of the domains, the unenforced would be TupleDomain.all().
+
+        // In all 3 cases shown above, the unenforced is not TupleDomain.none().
+        checkArgument(!unenforced.isNone());
+
+        Map<ColumnHandle, Domain> predicateDomains = predicate.getDomains().get();
+        Map<ColumnHandle, Domain> unenforcedDomains = unenforced.getDomains().get();
+        ImmutableMap.Builder<ColumnHandle, Domain> enforcedDomainsBuilder = ImmutableMap.builder();
+        for (Map.Entry<ColumnHandle, Domain> entry : predicateDomains.entrySet()) {
+            ColumnHandle predicateColumnHandle = entry.getKey();
+            if (unenforcedDomains.containsKey(predicateColumnHandle)) {
+                checkArgument(
+                        entry.getValue().equals(unenforcedDomains.get(predicateColumnHandle)),
+                        "Enforced tuple domain cannot be determined. The connector is expected to enforce the respective domain entirely on none, some, or all of the column.");
+            }
+            else {
+                enforcedDomainsBuilder.put(predicateColumnHandle, entry.getValue());
+            }
+        }
+        Map<ColumnHandle, Domain> enforcedDomains = enforcedDomainsBuilder.build();
+        checkArgument(
+                enforcedDomains.size() + unenforcedDomains.size() == predicateDomains.size(),
+                "Enforced tuple domain cannot be determined. Connector returned an unenforced TupleDomain that contains columns not in predicate.");
+        return TupleDomain.withColumnDomains(enforcedDomains);
     }
 }
